requirements.txt
streamlit>=1.52.0
pandas>=2.3.0
openpyxl>=3.1.0
pyodbc>=5.3.0
python-dotenv>=1.2.0
pymongo>=4.16.0
